Update CNAME
silentpost is moving! We're adopting a new gTLD and moving our home on the web
to silentpost.digital. silentpost.co was often confusing for clients--they
assumed we had just forgotten to add the 'm' at the end of our tld, so emails
would get lost, deadlines would be missed--it was bad all around. This new
domain will alleviate that problem.